# Improve performance when adding widgets