@level420 2016-07-04
Brilliant @johnspackman ! I stumbled last week while profiling my app with chrome debugger over qx.lang.Array.contains taking approx. 30% of the total time while initializing my large editor widget. I'll test this PR soon and report back.