@level420 2016-07-04
Yes! Initialization went down from 14.5sec down to 10sec! That's 30%!